% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CalcLikelihood.R
\name{CalcLikelihood}
\alias{CalcLikelihood}
\title{Calculate the log likelihood of a DTM given a topic model}
\usage{
CalcLikelihood(dtm, phi, theta)
}
\arguments{
\item{dtm}{= the document term matrix of class Matrix}

\item{phi}{= the phi matrix whose rows index topics and columns index words. The i, j entries are P(word_i | topic_j)}

\item{theta}{= the theta matrix whose rows index documents and columns index topics. The i, j entries are P(topic_i | document_j)}
}
\description{
This function takes a DTM, phi matrix (P(word|topic)), and a theta matrix (P(topic|document)) and
    returns a single value for the likelihood of the data given the model.
}
\examples{
myll <- CalcLikelihood(dtm=mydtm, phi=myphi, theta=mytheta)
}

