% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CalcLikelihoodC}
\alias{CalcLikelihoodC}
\title{Calculate the log likelihood of a DTM given a model using sequential C++ code}
\usage{
CalcLikelihoodC(dtm, phi, theta)
}
\arguments{
\item{dtm}{A document term matrix}

\item{phi}{= a topics X terms dimensional matrix where each entry is p(term|topic)}

\item{theta}{= a documents X topics dimensional matrix where each entry is p(topic|document)//'}
}
\description{
This function calculates the log likelihood for a topic model. Generally, this function (CalcLikelihoodC)
is called by the R function CalcLikelihood, which has options for paralellization. If you call CalcLikelihood but set
\code{parallel=FALSE}, then CalcLikelihood passes off to CalclikelihoodC completely.
}

