% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CalcSumSquares}
\alias{CalcSumSquares}
\title{Calculate the sum of square errors for documents}
\usage{
CalcSumSquares(dtm, phi, theta, ybar)
}
\arguments{
\item{dtm}{A document term matrix}

\item{phi}{= a topics X terms dimensional matrix where each entry is p(term|topic)}

\item{theta}{= a documents X topics dimensional matrix where each entry is p(topic|document)//'}
}
\description{
This function calculates a vector whose entries are SSE and SST, respectively.
It is desinged to work with the function CalcTopicModelR2. Generally, this function (CalcSumSquares)
should is not designed to be called by itself.
}

