% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_metrics.R
\name{CalcTopicModelR2}
\alias{CalcTopicModelR2}
\title{Calculate the R-squared of a topic model.}
\usage{
CalcTopicModelR2(dtm, phi, theta, cpus = 1, ...)
}
\arguments{
\item{dtm}{A documents by terms dimensional document term matrix of class
\code{dgCMatrix} or of class \code{matrix}.}

\item{phi}{A topics by terms dimensional matrix where each entry is
p(term_i |topic_j)}

\item{theta}{A documents by topics dimensional matrix where each entry
is p(topic_j|document_d)}

\item{cpus}{An integer of how many threads to use. Defaults to 1 and is
equivalent of setting \code{threads} in \link{mvrsquared}[calc_rsquared].}

\item{...}{Other arguments to be passed to
\code{\link[mvrsquared]{calc_rsquared}}. See note, below.}
}
\value{
Returns an object of class \code{numeric} representing the proportion
of variability in the data that is explained by the topic model.
}
\description{
Function to calculate R-squared for a topic model. 
This is a wrapper for \code{mvrsquared::calc_rsquared} with additional
checks in place to check for the inputs.
}
\examples{
# Load a pre-formatted dtm and topic model
data(nih_sample_dtm) 
data(nih_sample_topic_model)

# Get the R-squared of the model
r2 <- CalcTopicModelR2(dtm = nih_sample_dtm, 
                     phi = nih_sample_topic_model$phi, 
                     theta = nih_sample_topic_model$theta)


r2
}
