% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CountTerms.R
\name{CountTerms}
\alias{CountTerms}
\title{Summarize a document term matrix subject to document frequency of words}
\usage{
CountTerms(freq.mat, max.docs, min.docs)
}
\arguments{
\item{freq.mat}{A data frame that is output from the TermDocFreq() function.}

\item{max.docs}{What is the maximum document frequency of words to include in this summary?}

\item{min.docs}{What is the minimum document frequency of words to include in this summary?}
}
\description{
This function is designed to work with an in-development shiny app. It's overkill to use on its own but you can if you want to.
}
\examples{
term.freq <- TermDocFreq(dtm)
mycount <- CountTerms(freq.mat=term.freq, max.docs=nrow(dtm)/2, min.docs=5)
}

