% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DepluralizeDtm.R
\name{DepluralizeDtm}
\alias{DepluralizeDtm}
\title{Run the CorrectS function on columns of a document term matrix of class Matrix.}
\usage{
DepluralizeDtm(dtm)
}
\arguments{
\item{dtm}{A document term matrix of class \code{dgCMatrix} whose colnames are tokens}
}
\value{
Returns a document term matrix of class \code{dgCMatrix}. The columns index
the de-pluralized tokens of the input document term matrix. In other words,
there will generally be fewer columns in the returned matrix than the
input matrix
}
\description{
Turns pluralizations of words in the columns of a document term
matrix to their singular form. Then aggregates all columns that now have the
same token. See example below.
}
\examples{
myvec <- c("the quick brown fox eats chickens", "the slow gray fox eats the slow chicken", "look at my horse", "my horses are amazing")
names(myvec) <- paste("doc", 1:length(myvec), sep="_")

dtm <- Vec2Dtm(vec = myvec, min.n.gram = 1, max.n.gram = 1)

colnames(dtm)
[1] "amazing"  "brown"    "chicken"  "chickens" "eats"     "fox"      "gray"     "horse"    "horses"   "quick"    "slow"

dtm
4 x 11 sparse Matrix of class "dgCMatrix"
   [[ suppressing 11 column names ‘amazing’, ‘brown’, ‘chicken’ ... ]]

doc_1 . 1 . 1 1 1 . . . 1 .
doc_2 . . 1 . 1 1 1 . . . 2
doc_3 . . . . . . . 1 . . .
doc_4 1 . . . . . . . 1 . .

dtm_new <- DepluralizeDtm(dtm = dtm)

dtm_new
4 x 9 sparse Matrix of class "dgCMatrix"
      amazing brown eat fox gray quick slow chicken horse
doc_1       .     1   1   1    .     1    .       1     .
doc_2       .     .   1   1    1     .    2       1     .
doc_3       .     .   .   .    .     .    .       .     1
doc_4       1     .   .   .    .     .    .       .     1
}

