% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Dtm2DocsC}
\alias{Dtm2DocsC}
\title{Convert a DTM to a Character Vector of documents}
\usage{
Dtm2DocsC(dtm, vocab)
}
\arguments{
\item{dtm}{A sparse Matrix from the matrix package whose rownames correspond to documents and colnames correspond to words}

\item{parallel}{Do you want to parallelize this function using snowfall? Default is FALSE}

\item{cpus}{If parallel is TRUE, the number of threads to use. (Recommendation is 4, for memory's sake)}
}
\description{
This function takes a sparse matrix (DTM) as input and returns a character vector
whose length is equal to the number of rows of the input DTM.
}
\examples{
Dtm2Docs(dtm=mydtm, parallel=TRUE, cpus=4)
}

