% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_modeling_core.R
\name{fitted.lda_topic_model}
\alias{fitted.lda_topic_model}
\title{Get the fitted values for a Latent Dirichlet Allocation topic model}
\usage{
\method{fitted}{lda_topic_model}(object, ...)
}
\arguments{
\item{object}{a fitted object of class \code{lda_topic_model}}

\item{...}{other arguments.}
}
\value{
a matrix of class \code{dgCMatrix}
}
\description{
Obtains the fitted values for the specified LDA model.
}
\examples{
# load a premade topic model
data(nih_sample_topic_model)
# fit values using fitted()
fitted_values <- fitted(nih_sample_topic_model)
fitted_values
}
