% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/FormatRawLdaOutput.R
\name{FormatRawLdaOutput}
\alias{FormatRawLdaOutput}
\title{Format Raw Output from lda.collapsed.gibbs.sampler()}
\usage{
FormatRawLdaOutput(lda.result, docnames, smooth = TRUE)
}
\arguments{
\item{lda.result}{The list value returned by lda.collapsed.gibbs.sampler()}

\item{docnames}{A character vector giving the names of documents. This is generally rownames(dtm).}

\item{smooth}{Logical. Do you want to smooth your topic proportions so that there is a positive value for each term in each topic? Defaults to TRUE}
}
\description{
extracts outputs from LDA model estimated with "lda" package by Jonathan Chang
}
\examples{
lex <- ConvertDtm2Docs(dtm)
lex <- lexicalize(lex, vocab=colnames(dtm))
lda <- lda.collapsed.gibbs.sampler(documents = lex, K = 100, vocab = colnames(dtm), num.iterations=2000, alpha=0.1, eta=0.05)
lda <- FormatRawLdaOutput(lda.result=lda, docnames=rownames(dtm), smooth=TRUE)
}

