% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GetPhiPrime.R
\name{GetPhiPrime}
\alias{GetPhiPrime}
\title{Calculate phi.prime = P(topic|word)}
\usage{
GetPhiPrime(phi, theta)
}
\arguments{
\item{phi}{= the phi matrix whose rows index topics and columns index words. The i, j entries are P(word_i | topic_j)}

\item{theta}{= the theta matrix whose rows index documents and columns index topics. The i, j entries are P(topic_i | document_j)}
}
\description{
This function takes a phi matrix (P(word|topic)) and a theta matrix (P(topic|document)) and
returns phi.prime = P(topic|word). Phi.prime can be used for classifying new documents and for topic labels.
}
\examples{
lda$phi.prime <- GetPhiPrime(phi=lda$phi, theta=lda$theta)
}

