% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GetProbableTerms.R
\name{GetProbableTerms}
\alias{GetProbableTerms}
\title{Get cluster labels using a "more probable" method of terms}
\usage{
GetProbableTerms(docnames, dtm, p.terms = NULL)
}
\arguments{
\item{docnames}{= character vector of rownames of dtm for set of documents}

\item{dtm}{= a document term matrix, preferably a Matrix package sparse matrix}

\item{p.terms}{= if not NULL, a numeric vector representing the probabiliy of each term in the corpus whose names correspond to colnames(dtm).}
}
\description{
Function extracts probable terms from a set of documents. Probable here implies more probable than in a corpus overall
}
\details{
A numeric vector of the format p.terms. The entries of the vectors correspond
     to the difference in the probability of drawing a term from the set of documents given by
     docnames and the probability of drawing that term from the corpus overall (p.terms)
}
\examples{
mydocs <- rownames(lda$theta)[ lda$theta[ , 5 ] >= 0.25 ] # documents with a topic proportion of .25 or higher for topic 5

term.probs <- Matrix::colSums(dtm) / sum(Matrix::colSums(dtm))

GetProbableTerms(docnames=mydocs, dtm=dtm, p.terms=term.probs)
}

