% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/HellDist.R
\name{HellDist}
\alias{HellDist}
\title{Hellinger Distance}
\usage{
HellDist(p, q)
}
\arguments{
\item{p}{A numeric vector representing a probability distribution}

\item{q}{A numeric vector of length length(p) representing a probability distribuiton}
}
\description{
Calculates the Hellinger distances between two discrete probability distributions p and q
}
\examples{
x <- rnorm(n = 100, mean = 1, sd = 1)
y <- x^2
HellDist(p = x, q = y)r
}
\keyword{distance}
\keyword{functions}

