\name{Internals}
\alias{CalcLikelihoodC}
\alias{CalcSumSquares}
\alias{dtm_to_lexicon_c}
\alias{Dtm2DocsC}
\alias{fit_lda_c}
\alias{predict_lda_c}
\alias{Hellinger_cpp}
\alias{HellingerMat}
\alias{JSD_cpp}
\alias{JSDmat}
\title{Internal helper functions for \code{textmineR}}
\description{
These functions are internal helper functions for \code{textmineR}. They are not
designed to be called by users. Each of the functions here are C++ functions. 
There are corresponding R functions that call these that add additional functionality.
}


