% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/JSD.R
\name{JSD}
\alias{JSD}
\title{Jensen-Shannon Divergence}
\usage{
JSD(p, q)
}
\arguments{
\item{p}{A numeric vector representing a probability distribution}

\item{q}{A numeric vector of length length(p) representing a probability distribuiton}
}
\description{
This function calculates the Jensen Shannon Divergence for two probability vectors, p and q.
}
\examples{
x <- rnorm(n = 100, mean = 1, sd = 1)
y <- x^2
JSD(p = x, q = y)
}
\keyword{distance}
\keyword{functions}

