% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LabelTopics.R
\name{LabelTopics}
\alias{LabelTopics}
\title{Get some topic labels using a "more probable" method of terms}
\usage{
LabelTopics(assignments, dtm, M = 2)
}
\arguments{
\item{assignments}{= a documents X topics matrix similar to \code{theta}. This will work best if this matrix is sparse, with only a few non-zero topics per document.}

\item{dtm}{= a document term matrix, preferably a Matrix package sparse matrix}

\item{M}{= the number of n-gram labels you want to return. Defaults to 2}
}
\description{
Function calls \code{textmineR::GetProbableTerms()} with some rules to get topic labels. This function is in "super-ultra-mega alpha".
}
\examples{
assignments <- t(apply(theta, 1, function(x){
  x[ x < 0.05 ] <- 0
  x / sum(x)
}))

labels <- LabelTopics(assignments = assignments, dtm = dtm, M = 2)
}

