% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MakeSparseDTM.R
\name{MakeSparseDTM}
\alias{MakeSparseDTM}
\title{Convert a sparse simple triplet document term matrix (tm package) to a sparse Matrix object (Matrix package).}
\usage{
MakeSparseDTM(dtm)
}
\arguments{
\item{dtm}{A document term matrix of class DocumentTermMatrix (tm package)}
}
\description{
This function converts objects of class DocumentTermMatrix to an object of class Matrix. Document
term matrices are assumed to be of class Matrix for the majority of text mining functions in package \code{textmineR}.
}
\examples{
mydocs <- c("the quick brown fox", "the slow gray fox", "look at my horse", "my horse is amazing")
mycorp <- Corpus(VectorSource(mydocs))
mydtm <- DocumentTermMatrix(mycorp)
mydtm_sparse <- MakeSparseDTM(dtm=mydtm)
}

