% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NgramTokenizer.R
\name{NgramTokenizer}
\alias{NgramTokenizer}
\title{Get n-grams when creating a document term matrix}
\usage{
NgramTokenizer(min, max)
}
\arguments{
\item{min}{The minimum size of n for your n-grams}

\item{max}{The maximum size of n for your n-grams}
}
\description{
Function creates a function to create ngrams when converting a Corpus object to a DocumentTermMatrix object
}
\details{
For bigrams min=max=2. For bigrams and trigrams min=2, max=3
}
\examples{
ngrams <- NgramTokenizer(1, 3) # for unigrams, bigrams, and trigrams
myDTM <- DocumentTermMatrix(myCorp, control = list(tokenize = ngrams))
}

