% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PartitionTopticModelData.R
\name{PartitionTopticModelData}
\alias{PartitionTopticModelData}
\title{Function splits a DTM and theta by documents for parallelization}
\usage{
PartitionTopticModelData(dtm, theta, partitions = 8)
}
\arguments{
\item{dtm}{= a documents X terms dimensional document term matrix in
sparse format from the Matrix package or a regular R matrix.
 Will *not* work on DTMs from the tm package or simple triplet matrices from the slam package.}

\item{theta}{= a documents X topics dimensional matrix where each entry is p(topic|document)}

\item{partitions}{= number of threads over which you will parallelize.}
}
\description{
Function splits a DTM and theta by documents for parallelization.
This function is used by several other functions in \code{textmineR}. It
shouldn't need to be used on its own unless you're customizing a solution.
}
\examples{
data_divided <- PartitionTopticModelData(dtm=mydtm, theta=mytheta, partitions=8)
}

