% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentage_distribution.R
\name{percentage_distribution}
\alias{percentage_distribution}
\title{Counts Percentage Distribution of Group Variable}
\usage{
percentage_distribution(data, group_var, rows_n = NULL)
}
\arguments{
\item{data}{A tibble or data frame object}

\item{group_var}{The group in which one wishes to get the % and raw counts for, this could be something like sentiment, topic or brand.}

\item{rows_n}{State how many rows of data in total for if the number of rows to divide the group variable by differs from the number of rows in the `data` object provided.}
}
\value{
Prints the percentage distribution of any given variable
}
\description{
Counts Percentage Distribution of Group Variable
}
\details{
Shows the user the percentage distribution of a chosen group variable as well as the raw n counts for each.
}
\examples{
percentage_distribution(data = data,
                        group_var = group_var)
}
