% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_donut_chart.R
\name{plot_donut_chart}
\alias{plot_donut_chart}
\title{Plot a Donut Showing Group Distribution}
\usage{
plot_donut_chart(
  data,
  group_variable,
  custom_colours = NULL,
  legend_position = c("right", "left", "bottom", "none")[1],
  legend_text_size = 12,
  legend_title = "Group Distribution Donut Chart"
)
}
\arguments{
\item{data}{A dataframe or tibble object}

\item{group_variable}{The group in which the user wishes to display distribution}

\item{custom_colours}{For if the user has a customised colour pallatte to apply to the group variables. This should be supplied as a vector, if not the default colour palette is "Set3" of `scale_fill_brewer()`}

\item{legend_position}{The options are; "right", "left", "bottom" or "none", the default is "right".}

\item{legend_text_size}{For if the users wishes to change the size of the legend text, the default is set to size = 12.}

\item{legend_title}{The user may amend the title by upplying this as a character string.}
}
\value{
A ggplot2 visual, showing the distribution of a group variable byway of a donut chart.
}
\description{
Plot a Donut Showing Group Distribution
}
\examples{
theme_colours <- c("theme_1" = "#004d00",
                   "theme_2" = "#009900",
                   "theme_3" = "#ccffcc")
TextmineR::plot_donut_chart(data = data,
                            group_variable = theme,
                            custom_colours = theme_colours)
}
