% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_group_volume_over_time.R
\name{plot_group_volume_over_time}
\alias{plot_group_volume_over_time}
\title{Plot Data Volumes over Time per Group}
\usage{
plot_group_volume_over_time(
  data,
  date_variable,
  group_variable,
  group_colour = NULL,
  unit = c("1 day", "1 week", "1 month", "1 quarter", "1 year"),
  axis_text_unit = "1 month",
  rows = 2,
  fix_y_axis = TRUE
)
}
\arguments{
\item{data}{A tibble or data frame object}

\item{date_variable}{The column containing dates for each observation or row of data, must be of class Date}

\item{group_variable}{The group variable that the user wishes to compare volumes}

\item{group_colour}{A custom colour vector for if the user wishes to apply, the default pallet is ggplot2::scale_fill_brewer(palette = "Set3")}

\item{unit}{ow often should the dates be grouped when being summarised, the options are but not limited to; "1 day", "1 week", "1 month", "1 quarter", "1 year", for example; one may wish to call unit = "2 weeks"}

\item{axis_text_unit}{How often should the dates be plotted onto the x axis, the options are but not limited to; "1 day", "1 week", "1 month", "1 quarter", "1 year", for example; one may wish to call unit = "2 weeks"}

\item{rows}{How many rows the user wishes to plot the group volume over time charts across, the default is 2 rows}

\item{fix_y_axis}{A logical argument for if the user wishes to fix the y axis limit for each group based on the max value, the default is TRUE to display the comparison in group sizes.}
}
\value{
A ggplot2 visual showing the volumes over time of a given group within a dataset
}
\description{
#' @details
A function to render a simple volume over time chart but faceted for each level in a group variable, alike `plot_volume_over_time`, the user can change the unit, axis text breaks by unit, colours for each group and how many rows they wish to plot each group across. The user may also wish to fix the y axis limit, the default is TRUE to show volumes in comparison to each group more clearly.
}
\examples{
data \%>\%
 plot_group_volume_over_time(date_variable = date,
                             group_variable = brand,
                             group_colour = custom_colours,
                             unit = "1 month",
                             axis_text_unit = "1 year",
                             rows = 3,
                             fix_y_axis = TRUE)
}
