% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volume_over_time.R
\name{plot_volume_over_time}
\alias{plot_volume_over_time}
\title{Plot Data Volumes over Time}
\usage{
plot_volume_over_time(
  data,
  date_variable = date,
  smooth = TRUE,
  vline_date = NULL,
  bar_colour = "black",
  line_colour = "gold",
  unit = c("1 day", "1 week", "1 month", "1 quarter", "1 year"),
  axis_text_unit = "1 month"
)
}
\arguments{
\item{data}{A tibble or data frame object}

\item{date_variable}{The column containing dates for each observation or row of data, must be of class Date}

\item{smooth}{A logical argument stating whether to include a smooth trend line, the default is TRUE}

\item{vline_date}{For if the user wishes to plot a vertical line at any given date to point toward any changes or notable moments within the volume over time plot, argument expects a date given as a string, default option is NULL, no vertical line}

\item{bar_colour}{What colour should the volume bars be, given as a string and the default is "black"}

\item{line_colour}{What colour should the trend line be if smooth = TRUE, given as a string and the default is "yellow"}

\item{unit}{How often should the dates be grouped when being summarised, the options are but not limited to; "1 day", "1 week", "1 month", "1 quarter", "1 year", for example; one may wish to call unit = "2 weeks"}

\item{axis_text_unit}{How often should the dates be plotted onto the x axis, the options are but not limited to; "1 day", "1 week", "1 month", "1 quarter", "1 year", for example; one may wish to call unit = "2 weeks"}
}
\value{
A ggplot2 visual showing the volumes over time of a given dataset
}
\description{
Plot Data Volumes over Time
}
\details{
A function to render a simple volume over time chart, with the option to change the unit, axis text breaks by unit, colour and trend line options. The user may also include a dashed vertical line at a given date for if they wish.
}
\examples{
data \%>\%
 plot_volume_over_time(date_variable = date,
                       smooth = TRUE,
                       vline_date = NULL,
                       bar_colour = "black",
                       line_colour = "gold",
                       unit = "1 month",
                       axis_text_unit = "1 year")
}
