% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pos_udpipe_load_model.R
\name{pos_udpipe_load_model}
\alias{pos_udpipe_load_model}
\title{Load a Pre-trained Parts of Speech model from UDPipe}
\usage{
pos_udpipe_load_model(model)
}
\arguments{
\item{model}{The chosen model indicating what language the user wishes to select from udpipe's pre-trained parts of speech models. There are 65 options to choose from}
}
\value{
Loads the chosen UDPipe model into memory, ready for the annotation steps in the parts of speech workflow.
}
\description{
Load a Pre-trained Parts of Speech model from UDPipe
}
\details{
A function that wraps around the UDPipe package and retrieves/downloads pre-built models previously trained by the community, covering 65 different languages. For more information on what models are available, visit:[UDPipe Documentation](https://bnosac.github.io/udpipe/docs/doc1.html#pre-trained-models)
}
\examples{
pos_model_en <- TextmineR::pos_udpipe_load_model(model = "english")
}
