% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_modeling_core.R
\name{posterior}
\alias{posterior}
\title{Posterior methods for topic models}
\usage{
posterior(object, ...)
}
\arguments{
\item{object}{An existing trained topic model}

\item{...}{Additional arguments to the call}
}
\description{
\code{posterior} will draw from the posterior distribution of a
topic model
}
