% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_documents_add_line_breaks.R
\name{process_documents_add_line_breaks}
\alias{process_documents_add_line_breaks}
\title{Add Line Breaks to Documents}
\usage{
process_documents_add_line_breaks(text_variable, word_interval = 10)
}
\arguments{
\item{text_variable}{The column containing documents or texts the user wishes to add line breaks()}

\item{word_interval}{An integer expressing how many words after should each line break appear}
}
\value{
Additional line breaks to the chosen text variable.
}
\description{
Add Line Breaks to Documents
}
\examples{
data <- data \%>\% dplyr::mutate(texts_with_breaks = sapply(texts, process_documents_add_line_breaks))
}
