% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_documents_to_sentences.R
\name{process_documents_to_sentences}
\alias{process_documents_to_sentences}
\title{Break Documents out by Sentence}
\usage{
process_documents_to_sentences(data, text_variable)
}
\arguments{
\item{data}{A dataframe of tibble object containing documents}

\item{text_variable}{The column containing documents that the user wishes to unnest into sentences}
}
\value{
A dataframe with documents broken out by sentence, where each row represents a sentence.
}
\description{
Break Documents out by Sentence
}
\details{
Often in NLP, it's necessary to break documents up into sentences to perform lower-level analyses upon. This function serves as a means for doing so, outputting a dataframe containing each sentence beloning to documents.
}
\examples{
sentences_df <- process_documents_to_sentences(data = data,
                                               text_variable = texts)
}
