% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecursiveRbind.R
\name{RecursiveRbind}
\alias{RecursiveRbind}
\title{Recursively call rBind from the Matrix package.}
\usage{
RecursiveRbind(matrix_list)
}
\arguments{
\item{matrix_list}{A list, each element containing a matrix of class \code{dgCMatrix}}
}
\value{
Returns a matrix of class \code{dgCMatrix}.
}
\description{
This is used for combining a list of sparce matrices into a large Matrix.
This is a worker function for \code{textmineR} and generally not made to be
used by users. However, the function is provided for custom functions.
}
\examples{
a_matrix <- Matrix::Matrix(0, nrow=10, ncol=10)

a_list <- list(a_matrix, a_matrix, a_matrix, a_matrix, a_matrix)

result <- RecursiveRbind(a_list)

}

