% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SubsetDtm.R
\name{SubsetDtm}
\alias{SubsetDtm}
\title{Subset a document term matrix}
\usage{
SubsetDtm(dtm, keep.terms)
}
\arguments{
\item{dtm}{A document term matrix of class Matrix}

\item{keep.terms}{A character vector of colnames(dtm) indicating which terms you want to keep.}
}
\description{
This function is designed to work with an in-development shiny app. It's overkill to use on its own but you can if you want to...
}

