% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmineR-deprecated.R
\name{textmineR-deprecated}
\alias{textmineR-deprecated}
\title{Deprecated functions in package \pkg{textmineR}.}
\description{
The functions listed below are deprecated and will be defunct in
  the near future. When possible, alternative functions with similar
  functionality are also mentioned.
}
\details{
Below is a list of deprecated functions:
RecursiveRbind
}
\keyword{internal}
