% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TopicModelR2.R
\name{TopicModelR2}
\alias{TopicModelR2}
\title{Function to calculate R-squared of a topic model.}
\usage{
TopicModelR2(dtm, phi, theta, parallel = FALSE, cpus = NULL)
}
\arguments{
\item{dtm}{= a documents X terms dimensional document term matrix in
sparse format from the Matrix package or a regular R matrix.
 Will *not* work on DTMs from the tm package or simple triplet matrices from the slam package.}

\item{phi}{= a topics X terms dimensional matrix where each entry is p(term|topic)}

\item{theta}{= a documents X topics dimensional matrix where each entry is p(topic|document)}

\item{parallel}{= a logical. Do you have snowfall installed? Would you like to parallelize?}

\item{cpus}{= number of threads over which to parallelize.}
}
\description{
Function to calculate R-squared for a topic model using pure R code.
Note: This function is slower than CalcTopicModelR2, which is a C++ function.
However, TopicModelR2 gives you more output, such as the sse, ssm, and sst for each document.
This uses a geometric interpretation of R-squared as the proportion of total distance
each document is from the center of all the documents that is explained by the model.
Note: all input matrices must have rownames and colnames
}
\details{
Output:
a list with 3 elements -
  r2 = R-squared of the model
  sse = the sum of squared errors. This is a vector, the square root of which
      gives the l2-norm or euclidean distance from each document to its fitted value
  sst = the total sum of squares. This is a vector, the square root of which
      gives the l2-norm or euclidean distance from each document to the "mean" document.
  ssm = the model sum of squares. This is a vector, the square root of which
      gives the l2-norm or euclidean distance from the fitted-value for each document to the "mean" document.
}
\examples{
r2 <- TopicModelR2(dtm=mydtm, phi=lda$phi, theta=lda$theta, parallel=TRUE, cpus=8)
}

