% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_modeling_core.R
\name{update}
\alias{update}
\title{Update methods for topic models}
\usage{
update(object, ...)
}
\arguments{
\item{object}{An existing trained topic model}

\item{...}{Additional arguments to the call}
}
\description{
\code{update} will update a previously-trained topic model based
on new data. Useful for updates or transfer learning.
}
