% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CalcTopicModelR2.R
\name{CalcTopicModelR2}
\alias{CalcTopicModelR2}
\title{Function to calculate R-squared of a topic model.}
\usage{
CalcTopicModelR2(dtm, phi, theta)
}
\arguments{
\item{dtm}{= a documents X terms dimensional document term matrix in
sparse format from the Matrix package or a regular R matrix.
 Will *not* work on DTMs from the tm package or simple triplet matrices from the slam package.}

\item{phi}{= a topics X terms dimensional matrix where each entry is p(term|topic)}

\item{theta}{= a documents X topics dimensional matrix where each entry is p(topic|document)}

\item{parallel}{= a logical. Do you have snowfall installed? Would you like to parallelize?}

\item{cpus}{= number of threads over which to parallelize.}
}
\description{
Function to calculate R-squared for a topic model.
This uses a geometric interpretation of R-squared as the proportion of total distance
each document is from the center of all the documents that is explained by the model.
}
\examples{
r2 <- CalcTopicModelR2(dtm=mydtm, phi=lda$phi, theta=lda$theta, parallel=TRUE, cpus=8)
}

