% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Files2Vec.R
\name{Files2Vec}
\alias{Files2Vec}
\title{Function for reading text files into R}
\usage{
Files2Vec(directory, ...)
}
\arguments{
\item{directory}{A path to directory containing the files you want to read into R.}

\item{...}{Other arguments to be passed to \code{\link[textmineR]{TmParallelApply}}. See note, below.}
}
\value{
Returns a character vector where each entry corresponds to a document.
}
\description{
This function reads in all files in a directory ending in .txt into R.
The result is a character vector where each entry is a .txt file.
The names of the resulting vector are derived from the names of the files.
}
\note{
This function performs parallel computation by default. The default 
behavior is to use all available cores according to \code{\link[parallel]{detectCores}}.
However, this can be modified by passing the \code{cpus} argument when calling
this function.
}
\examples{
\dontrun{
my_text_vector <- Files2Vec("/path/to/my/data/")
}
}

