% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitCtmModel.R
\name{FitCtmModel}
\alias{FitCtmModel}
\title{Fit a Correlated Topic Model}
\usage{
FitCtmModel(dtm, k, return_all = TRUE, ...)
}
\arguments{
\item{dtm}{A document term matrix of class \code{dgCMatrix}}

\item{k}{Number of topics}

\item{return_all}{Logical. Do you want the raw results of the underlying 
function returned along with the formatted results? Defaults to \code{TRUE}.}

\item{...}{Other arguments to pass to \link[topicmodels]{CTM}.}
}
\value{
Returns a list with a minumum of two objects, \code{phi} and 
\code{theta}. The rows of \code{phi} index topics and the columns index tokens.
The rows of \code{theta} index documents and the columns index topics.
}
\description{
A wrapper for the \link[topicmodels]{CTM} function based on 
Blei's original code that returns a nicely-formatted topic model.
}
\examples{
# Load a pre-formatted dtm 
data(nih_sample_dtm) 

# Fit a CTM model on a sample of documents
model <- FitCtmModel(dtm = nih_sample_dtm[ sample(1:nrow(nih_sample_dtm) , 10) , ], 
                     k = 3, return_all = FALSE)
}
