% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FitLdaModel.R
\name{FitLdaModel}
\alias{FitLdaModel}
\title{Fit a topic model using Latent Dirichlet Allocation}
\usage{
FitLdaModel(dtm, k, iterations, alpha = 0.1, beta = 0.05, ...)
}
\arguments{
\item{dtm}{A document term matrix of class \code{Matrix::dgCMatrix}}

\item{k}{Number of topics}

\item{iterations}{The number of Gibbs iterations}

\item{alpha}{Dirichlet parameter for the distribution of topics over documents. Defaults to 0.1}

\item{beta}{Dirichlet parameter for the distribution of words over topics. Defaults to 0.05}

\item{...}{Other arguments to pass to \code{lda::lda.collapsed.gibbs.sampler}.}
}
\description{
A wrapper for \code{lda::lda.collapsed.gibbs.sampler} that returns nicely formatted objects.
}
\examples{
# fit a model with the default options
model <- FitLdaModel(dtm = mydtm, iterations = 2000)

# include likelihoods
model <- FitLdaModel(dtm = mydtm, iterations = 2000, compute.log.likelihood = T)
}

