% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_modeling_core.R
\name{FitLdaModel}
\alias{FitLdaModel}
\title{Fit a Latent Dirichlet Allocation topic model}
\usage{
FitLdaModel(dtm, k, iterations = NULL, burnin = -1, alpha = 0.1,
  beta = 0.05, optimize_alpha = FALSE, calc_likelihood = FALSE,
  calc_coherence = TRUE, calc_r2 = FALSE, seed = NULL, ...)
}
\arguments{
\item{dtm}{A document term matrix or term co-occurrence matrix of class dgCMatrix}

\item{k}{Integer number of topics}

\item{iterations}{Integer number of iterations for the Gibbs sampler to run. A
future version may include automatic stopping criteria.}

\item{alpha}{Vector of length \code{k} for asymmetric or a number for symmetric.
This is the prior for topics over documents}

\item{beta}{Vector of length \code{ncol(dtm)} for asymmetric or a number for symmetric.
This is the prior for words over topics.}

\item{seed}{If not null (the default) then the random seed you wish to set. This
will return the same outputs for the same inputs. Useful for diagnostics.}

\item{...}{Other arguments to be passed to textmineR::TmParallelApply}
}
\value{
Returns an S3 object of class c("LDA", "TopicModel"). DESCRIBE MORE
}
\description{
Fit a Latent Dirichlet Allocation topic model using collapsed Gibbs sampling.
}
\details{
EXPLAIN IMPLEMENTATION DETAILS
}
\examples{
GIVE EXAMPLES
}
