% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GetTopTerms.R
\name{GetTopTerms}
\alias{GetTopTerms}
\title{Get Top Terms for each topic from a topic model}
\usage{
GetTopTerms(phi, M)
}
\arguments{
\item{phi}{A matrix whose rows index topics and columns index words}

\item{M}{An integer for the number of terms to return}
}
\description{
Takes topics X terms matrix and returns top M terms for each topic
}
\examples{
GetTopTerms(phi=my_phi, M=5)
}

