% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NgramTokenizer.R
\name{NgramTokenizer}
\alias{NgramTokenizer}
\title{Get n-grams when creating a document term matrix}
\usage{
NgramTokenizer(min, max)
}
\arguments{
\item{min}{Smallest n-gram desired}

\item{max}{Largest n-gram desired}
}
\value{
Returns a function
}
\description{
This function is a helper function for \code{textmineR}. It is
not designed for users. It aids in creating n-grams in DTM construction
}
\examples{
\dontrun{
my_tokenizer <- NgramTokenizer(min=1, max=3)

dtm <- tm::DocumentTermMatrix(corp, control=list(tokenize=my_tokenizer))

dtm <- MakeSparseDTM(dtm)
}
}

