% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TermDocFreq.R
\name{TermDocFreq}
\alias{TermDocFreq}
\title{Get term frequencies and document frequencies from a document term matrix.}
\usage{
TermDocFreq(dtm)
}
\arguments{
\item{dtm}{A document term matrix of class \code{dgCMatrix}.}
}
\value{
Returns a \code{data.frame} with 4 columns. The first column,
\code{term} is a vector of token labels. The second column, \code{term.freq}
is the count of times \code{term} appears in the entire corpus. The third
column \code{doc.freq} is the count of the number of documents in which
\code{term} appears. The fourth column, \code{idf} is the log-weighted
inverse document frequency of \code{term}.
}
\description{
This function takes a document term matrix as input and
returns a data frame with columns for term frequency, document frequency,
and inverse-document frequency
}
\examples{
\dontrun{
myvec <- c("the quick brown fox eats chickens",
           "the slow gray fox eats the slow chicken",
           "look at my horse", "my horses are amazing")
names(myvec) <- paste("doc", 1:length(myvec), sep="_")

dtm <- Vec2Dtm(vec = myvec, min.n.gram = 1, max.n.gram = 1)

term_freq_mat <- TermDocFreq(dtm)
}
}

