% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmineR.R
\docType{package}
\name{textmineR}
\alias{textmineR}
\alias{textmineR-package}
\title{textmineR}
\description{
Functions for Text Mining and Topic Modeling
}
\details{
An aid for text mining in R, with a syntax that
should be familiar to experienced R users. Provides a wrapper for several 
topic models that take similarly-formatted input and give similarly-formatted
output. Has additional functionality for analyzing and diagnostics for
topic models.
}
